requirements.txt
# Web framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# AI / Groq
groq>=0.12.0

# Audio processing
pydub>=0.25.1

# RSS parsing
feedparser>=6.0.0

# HTTP client
httpx>=0.27.0

# Azure
azure-storage-blob>=12.20.0
aiohttp>=3.9.0

# Configuration
pydantic-settings>=2.5.0

# Scheduling
apscheduler>=3.10.0

# MCP
fastmcp>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
